requirements.txt
Django==1.11.14
djangorestframework==3.8.2